+++ packages/adapter-node/rollup.config.js
@@ -1,24 +1,13 @@
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import typescript from 'rollup-plugin-typescript2';
 
-export default [
-	{
-		input: 'src/index.ts',
-		output: {
-			file: 'index.js',
-			format: 'cjs'
-		},
-		plugins: [nodeResolve(), commonjs(), typescript()],
-		external: require('module').builtinModules
+export default {
+	input: 'src/server.js',
+	output: {
+		file: 'files/server.js',
+		format: 'cjs',
+		sourcemap: true
 	},
-	{
-		input: 'src/server.ts',
-		output: {
-			file: 'server.js',
-			format: 'cjs'
-		},
-		plugins: [nodeResolve(), commonjs(), typescript()],
-		external: require('module').builtinModules
-	}
-];
+	plugins: [nodeResolve(), commonjs()],
+	external: require('module').builtinModules
+};